// src/model/interface.ts
export interface IData {
  id: number
  company: string
  title: string
  description: string
  price: number
  discount: number
  images: string[]
}

export interface ICart {
  id: number
  company: string
  title: string
  description: string
  price: number
  discount: number
  images: string[]
  count: number
}